requirements.txt
discord.py>=1.6.0
requests-oauthlib>=1.3.0
tornado==6.1
mysql-connector-python==8.0.23
